Homescreen widget display modifications
Main column in homescreen is now wrapped in a SingleChildScrollView because this allows the content to be scrolled when on-screen keyboard is active. This also gets rid of the previous render overflow errors caused by nested listviews and such.

Upcoming forecast display now loops through the list and return a Column populated by the list items, instead of using the previous ListView.builder which was causing problems because of nesting